requirements.txt
streamlit
supabase
sqlmodel
PyJWT
streamlit-local-storage